fix(CAT-23): BEM and game-over name
