+++ packages/client/package.json
@@ -17,7 +17,7 @@
     "eslint-config-prettier": "8.5.0",
     "prettier": "2.7.1",
     "react": "18.2.0",
-    "react-confetti": "^6.1.0",
+    "react-confetti": "6.1.0",
     "react-dom": "18.2.0",
     "react-router-dom": "6.22.3"
   },
