+++ packages/client/src/components/exit-button/exit-button.css
@@ -1,19 +1,22 @@
-.exit__container {
+.exit-button {
   display: flex;
   justify-content: flex-end;
   margin-top: calc(100vh - 600px);
 }
 
-.exit__link {
+.exit-button__button {
   display: flex;
   justify-content: flex-end;
   flex-direction: row;
   gap: 12px;
   align-items: center;
   text-decoration: none;
+  background-color: #efe5cc;
+  border: none;
+  cursor: pointer;
 }
 
-.exit__text {
-  font-size: 12px;
+.exit-button__button-text {
+  font-size: 16px;
   color: #565a5d;
 }
