+++ packages/client/src/components/exit-button/index.ts
@@ -1,0 +1,1 @@
+export { ExitButtom } from './exit-button';
