--- packages/client/src/components/exit/index.ts
@@ -1,1 +1,0 @@
-export { Exit } from './exit';
