+++ packages/client/src/components/game-over-wrapper/index.ts
@@ -1,0 +1,1 @@
+export { GameOverWrapper } from './game-over-wrapper';
