--- packages/client/src/components/game-over/index.ts
@@ -1,1 +1,0 @@
-export { GameOver } from './game-over';
