+++ packages/client/src/components/index.ts
@@ -1,4 +1,4 @@
 export { PageWrapper } from './page-wrapper';
 export { NavigationBar } from './navigation-bar';
 export { IconButton } from './icon-button';
-export { GameOver } from './game-over';
+export { GameOverWrapper } from './game-over-wrapper';
